Interim commit in RAPTOR development cycle

+++ src/PublicTransportRouter/GTFSDataManager/Transfer.java
@@ -3,12 +3,12 @@ package src.PublicTransportRouter.GTFSDataManager;
 import java.util.HashMap;
 
 public class Transfer {
-    private HashMap<String, Integer> transferMap;
-    /* Strings refer to destination stops' IDs, and integers refer to minimum transfer times; origin stop IDs are keys
+    private HashMap<String, Double> transferMap;
+    /* Strings refer to destination stops' IDs, and doubles refer to minimum transfer times; origin stop IDs are keys
     of the pertinent hashmap
     */
 
-    Transfer(HashMap<String, Integer> transferMap) {
+    Transfer(HashMap<String, Double> transferMap) {
         this.transferMap = transferMap;
     }
 
@@ -16,7 +16,7 @@ public class Transfer {
         this(new HashMap<>());    // All values are initialized to nulls and zeroes
     }
 
-    public HashMap<String, Integer> getTransferMap() {
+    public HashMap<String, Double> getTransferMap() {
         return this.transferMap;
     }
 }
